Move all Topics configuration objects into subpackage named topics
