+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/services/asn1/message/AsnCodecMessageServiceController.java
@@ -3,9 +3,9 @@ package us.dot.its.jpo.ode.services.asn1.message;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import us.dot.its.jpo.ode.kafka.Asn1CoderTopics;
+import us.dot.its.jpo.ode.kafka.topics.Asn1CoderTopics;
 import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
-import us.dot.its.jpo.ode.kafka.RawEncodedJsonTopics;
+import us.dot.its.jpo.ode.kafka.topics.RawEncodedJsonTopics;
 import us.dot.its.jpo.ode.wrapper.MessageConsumer;
 
 /***
