+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/traveler/TimDepositController.java
@@ -23,10 +23,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import us.dot.its.jpo.ode.context.AppContext;
-import us.dot.its.jpo.ode.kafka.Asn1CoderTopics;
-import us.dot.its.jpo.ode.kafka.JsonTopics;
+import us.dot.its.jpo.ode.kafka.topics.Asn1CoderTopics;
+import us.dot.its.jpo.ode.kafka.topics.JsonTopics;
 import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
-import us.dot.its.jpo.ode.kafka.PojoTopics;
+import us.dot.its.jpo.ode.kafka.topics.PojoTopics;
 import us.dot.its.jpo.ode.model.OdeMsgMetadata.GeneratedBy;
 import us.dot.its.jpo.ode.model.*;
 import us.dot.its.jpo.ode.plugin.ServiceRequest;
@@ -44,7 +44,6 @@ import us.dot.its.jpo.ode.util.JsonUtils.JsonUtilsException;
 import us.dot.its.jpo.ode.util.XmlUtils;
 import us.dot.its.jpo.ode.wrapper.MessageProducer;
 import us.dot.its.jpo.ode.wrapper.serdes.OdeTimSerializer;
-import java.util.UUID;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
