+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/services/asn1/AsnCodecRouterServiceControllerTest.java
@@ -20,6 +20,10 @@ import mockit.Injectable;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import us.dot.its.jpo.ode.kafka.*;
+import us.dot.its.jpo.ode.kafka.topics.Asn1CoderTopics;
+import us.dot.its.jpo.ode.kafka.topics.JsonTopics;
+import us.dot.its.jpo.ode.kafka.topics.PojoTopics;
+import us.dot.its.jpo.ode.kafka.topics.SDXDepositorTopics;
 import us.dot.its.jpo.ode.rsu.RsuProperties;
 import us.dot.its.jpo.ode.security.SecurityServicesProperties;
 import us.dot.its.jpo.ode.wrapper.MessageConsumer;
