+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/services/json/ToJsonServiceControllerTest.java
@@ -19,9 +19,9 @@ import org.junit.jupiter.api.Test;
 
 import mockit.Expectations;
 import mockit.Injectable;
-import us.dot.its.jpo.ode.kafka.JsonTopics;
+import us.dot.its.jpo.ode.kafka.topics.JsonTopics;
 import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
-import us.dot.its.jpo.ode.kafka.PojoTopics;
+import us.dot.its.jpo.ode.kafka.topics.PojoTopics;
 import us.dot.its.jpo.ode.wrapper.MessageConsumer;
 import us.dot.its.jpo.ode.wrapper.MessageProcessor;
